+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Throw an exception if the `setter` method does not exists on entity
+- Leverage [Symfony Property Access](https://symfony.com/doc/current/components/property_access.html) component instead of homemade `$setter` and `$getter` methods (#92)
+- Throw an exception if the `setter` method does not exists on entity (#91)
 
 ## 0.31.1